html thymeleaf: index visible- html et css basique- tentative de associer le mois courant à la photo du mois avec un fichier image.properties pas concluant

+++ kakarot/src/main/java/repository/PlanteRepository.java
@@ -9,7 +9,7 @@ import entities.Plantes;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface PlanteRepository extends CrudRepository<Plantes, Long> {
+public interface PlanteRepository extends CrudRepository<Plantes, Integer> {
 
     Plantes findByName(String name);
